tests/wpt/web-platform-tests/html/semantics/forms/the-button-element/active-onblur.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" title="Joey Arhar" href="mailto:jarhar@chromium.org">
<link rel="help" href="http://crbug.com/945854">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<!-- This behavior is not explicitly specified. -->

<button id=b1>button one</button>
<button id=b2>button two</button>

<script>
promise_test(async () => {
  b1.focus();

  // Hold spacebar down
  await (new test_driver.Actions()).keyDown('\uE00D').send();
  assert_equals(document.querySelector(':active'), b1,
    'Buttons should be :active while the spacebar is pressed down.');

  // Press tab
  await (new test_driver.Actions()).keyDown('\uE004').keyUp('\uE004').send();
  assert_equals(document.querySelector(':active'), null,
    'Buttons should not be :active after tab is used to change focus.');

  // Release spacebar
  await (new test_driver.Actions()).keyUp('\uE00D').send();
}, 'Buttons should clear :active when the user tabs away from them while holding spacebar.');
</script>
